refactor : model파일명을 domain 파일명으로 변경
