+++ __tests__/Car.test.js
@@ -1,4 +1,4 @@
-import Car from "../src/model/Car.js";
+import Car from "../src/domain/Car.js";
 describe("차 객체 유닛 테스트", () => {
   test("차 객체를 만들면 차가 생성되어야 함", () => {
     //given
@@ -17,6 +17,7 @@ describe("차 객체 유닛 테스트", () => {
     const car = new Car("마빈");
     const initialPosition = 0;
     const EXPECTED_RESULT = initialPosition + 2;
+
     //when
     car.goForward();
     car.goForward();
