+++ src/domain/RandomNumberGenerator.js
