+++ src/service/raceService.js
@@ -1,8 +1,8 @@
-import Car from "../model/Car.js";
+import Car from "../domain/Car.js";
 import { InputView } from "../view/InputView.js";
 import { OutputView } from "../view/OutputView.js";
 import { parsingService } from "./parsingService.js";
-import randomNumberGenerator from "../model/RandomNumberGenerator.js";
+import randomNumberGenerator from "../domain/RandomNumberGenerator.js";
 import { systemSetting } from "../constants/systemSetting.js";
 
 export async function raceInit() {
